tests: run race and cover only in travis

+++ test.sh
@@ -12,11 +12,16 @@ echo ">> Generating Person ORM..."
 go generate ./example/...
 
 echo ">> Running tests..."
-echo "" > coverage.txt
-for d in $(go list ./... | grep -v vendor); do
-    go test -v -race -coverprofile=profile.out -covermode=atomic $d
-    if [ -f profile.out ]; then
-        cat profile.out >> coverage.txt
-        rm profile.out
-    fi
-done
+
+if [ -z "${TRAVIS}" ]; then
+	go test ./...
+else
+	echo "" > coverage.txt
+	for d in $(go list ./... | grep -v vendor); do
+		go test -v -race -coverprofile=profile.out -covermode=atomic $d
+		if [ -f profile.out ]; then
+			cat profile.out >> coverage.txt
+			rm profile.out
+		fi
+	done
+fi
